terminando el reto

+++ src/controller/products.controller.js
@@ -64,14 +64,29 @@ export class ProductsController {
     });
   };
 
-  static delete = async (req, res) => {
-    const id = req.params.productId;
-    const result = await ProductsService.deleteProduct(id);
-    logger.info("Producto eliminado exitosamente");
-    res.json({
-      status: "success",
-      data: result,
-      message: "Producto borrado correctamente",
-    });
+  static deleteProduct = async (req, res) => {
+    try {
+      const id = req.params.productId;
+      const product = await ProductsService.getProductById(id);
+      if (
+        (req.user.rol === "premium" && product.owner === req.user._id) ||
+        req.user.rol === "admin"
+      ) {
+        const result = await ProductsService.deleteProduct(id);
+        logger.info("Producto eliminado exitosamente");
+        res.json({
+          status: "success",
+          data: result,
+          message: "Producto borrado correctamente",
+        });
+      } else {
+        res.json({
+          status: "error",
+          message: "No tienes permiso para borrar este producto",
+        });
+      }
+    } catch (error) {
+      res.status.json({ status: "error", message: error.message });
+    }
   };
 }
